fix: rudderstack events
