+++ packages/bot-web-ui/src/analytics/rudderstack-quick-strategy.ts
@@ -2,30 +2,6 @@ import { Analytics, TEvents } from '@deriv-com/analytics';
 import { ACTION, form_name, type TFormStrategy } from './constants';
 import { getRsStrategyType, getSubpageName, getTradeParameterData } from './utils';
 
-export const rudderStackSendQsOpenEvent = ({ subform_source }: TEvents['ce_bot_form']) => {
-    Analytics.trackEvent('ce_bot_form', {
-        action: ACTION.OPEN,
-        form_name,
-        subpage_name: getSubpageName(),
-        subform_name: 'quick_strategy',
-        subform_source,
-    });
-};
-
-export const rudderStackSendQsCloseEvent = ({
-    quick_strategy_tab,
-    selected_strategy,
-}: TEvents['ce_bot_form'] & TFormStrategy) => {
-    Analytics.trackEvent('ce_bot_form', {
-        action: ACTION.CLOSE,
-        form_name,
-        subpage_name: getSubpageName(),
-        subform_name: 'quick_strategy',
-        quick_strategy_tab,
-        strategy_name: getRsStrategyType(selected_strategy),
-    });
-};
-
 export const rudderStackSendQsRunStrategyEvent = ({
     form_values,
     selected_strategy,
