+++ packages/bot-web-ui/src/pages/bot-builder/toolbar/workspace-group.tsx
@@ -3,7 +3,7 @@ import { DesktopWrapper } from '@deriv/components';
 import { observer } from '@deriv/stores';
 import { localize } from '@deriv/translations';
 import { useDBotStore } from 'Stores/useDBotStore';
-import { rudderStackSendSwitchLoadStrategyTabEvent } from '../../../analytics/rudderstack-bot-builder';
+import { rudderStackSendOpenEvent } from '../../../analytics/rudderstack-common-events';
 import ToolbarIcon from './toolbar-icon';
 
 const WorkspaceGroup = observer(() => {
@@ -31,7 +31,7 @@ const WorkspaceGroup = observer(() => {
                     action={() => {
                         setPreviewOnPopup(true);
                         toggleLoadModal();
-                        rudderStackSendSwitchLoadStrategyTabEvent({ load_strategy_tab: 'recent' });
+                        rudderStackSendOpenEvent({ subform_source: 'bot_builder', subform_name: 'load_strategy' });
                     }}
                 />
                 <ToolbarIcon
